feat: moved main function into lib.rs to allow wasm builds
